# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: HalalChain V2 API
  description: Enterprise Halal Supply Chain Platform — REST API
  version: "2.0.0"
  contact:
    name: HalalChain Team
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.halalchain.app
    description: Production server

# ─────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  # ─────────────────────────────────────────────────────────────
  # Schemas
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    Supplier:
      type: object
      required:
        - id
        - name
        - country
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: "Al-Barakah Foods Sdn Bhd"
        country:
          type: string
          example: "Malaysia"
        email:
          type: string
          format: email
          nullable: true
          example: "contact@albarakah.com"
        phone:
          type: string
          nullable: true
          example: "+60123456789"
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-10T06:00:00.000Z"

    # ── Product ────────────────────────────────────────────────
...
    Shipment:
      type: object
      required:
        - id
        - purchaseOrderId
        - trackingNumber
        - origin
        - destination
        - status
      properties:
        id:
          type: string
          format: uuid
          example: "ee0e8400-e29b-41d4-a716-446655440009"
        purchaseOrderId:
          type: string
          format: uuid
          example: "cc0e8400-e29b-41d4-a716-446655440007"
        trackingNumber:
          type: string
          example: "TRK-2024-98765"
        origin:
          type: string
          example: "Kuala Lumpur, Malaysia"
        destination:
          type: string
          example: "Jakarta, Indonesia"
        status:
          type: string
          enum:
            - PENDING
            - IN_TRANSIT
            - DELIVERED
            - DELAYED
          example: "IN_TRANSIT"
        estimatedArrival:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-10T14:00:00.000Z"
        shippedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-02T06:00:00.000Z"

    # ── AuditLog ───────────────────────────────────────────────
...
    PaginatedResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
        - totalPages
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of result items for the current page
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 85
        totalPages:
          type: integer
          minimum: 0
          example: 5

    # ── ErrorResponse ──────────────────────────────────────────
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Resource not found"

# ─────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────
paths:
...
  /api/suppliers:
    get:
      tags:
        - Suppliers
      summary: List suppliers (paginated, filterable)
      description: Returns a paginated list of suppliers. Supports filtering by name, country, and status. Requires ADMIN or MANAGER role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (default 1, minimum 1)
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Items per page (default 20, maximum 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: name
          in: query
          description: Filter by supplier name (case-insensitive, partial match)
          schema:
            type: string
            example: "Al-Barakah"
        - name: country
          in: query
          description: Filter by country (case-insensitive, exact match)
          schema:
            type: string
            example: "Malaysia"
        - name: status
          in: query
          description: Filter by supplier status
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
            example: "ACTIVE"
      responses:
        "200":
          description: Paginated list of suppliers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Supplier"
              example:
                data:
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "Al-Barakah Foods Sdn Bhd"
                    country: "Malaysia"
                    email: "contact@albarakah.com"
                    phone: "+60123456789"
                    status: "ACTIVE"
                    createdAt: "2024-01-10T06:00:00.000Z"
                  - id: "661e8400-e29b-41d4-a716-446655440002"
                    name: "Halal Fresh Indonesia"
                    country: "Indonesia"
                    email: "info@halalfresh.id"
                    phone: "+62217654321"
                    status: "ACTIVE"
                    createdAt: "2024-02-15T09:30:00.000Z"
                page: 1
                limit: 20
                total: 2
                totalPages: 1
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Suppliers
      summary: Create a new supplier (ADMIN only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - country
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                  example: "Halal Fresh Indonesia"
                country:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Indonesia"
                email:
                  type: string
                  format: email
                  example: "info@halalfresh.id"
                phone:
                  type: string
                  maxLength: 30
                  example: "+62217654321"
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
                  example: "ACTIVE"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                type: object
                properties:
                  supplier:
                    $ref: "#/components/schemas/Supplier"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/suppliers/{id}:
    get:
      tags:
        - Suppliers
      summary: Get a single supplier by ID
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Supplier details including products and certificates
          content:
            application/json:
              schema:
                type: object
                properties:
                  supplier:
                    $ref: "#/components/schemas/Supplier"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Supplier not found"

    patch:
      tags:
        - Suppliers
      summary: Update a supplier (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                country:
                  type: string
                  minLength: 2
                  maxLength: 100
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  maxLength: 30
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  supplier:
                    $ref: "#/components/schemas/Supplier"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Suppliers
      summary: Delete a supplier (ADMIN only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Supplier deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Supplier deleted"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ══════════════════════════════════════════════════════════
  # Products
  # ══════════════════════════════════════════════════════════
...
  /api/shipments:
    get:
      tags:
        - Shipments
      summary: List shipments (paginated)
      description: Returns a paginated list of shipments. Requires ADMIN or MANAGER role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of shipments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Shipment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/shipments/{id}:
    get:
      tags:
        - Shipments
      summary: Get a single shipment by ID
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Shipment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipment:
                    $ref: "#/components/schemas/Shipment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Shipment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...